Add files via upload
Heres the "grid". the important bits here are gridManager, set up to spawn any gameobject on every tile with a sprite on it on a selected tilemap, and the tile object which is an empty object with a correctly sized collider. Player movement here should be fairly straightforward using the colliders to select tiles and setting position to them, and the "occupant" variable on the tiles themselves can probably be ignored for now.

 most of the time i spent on this this saturday wasn't on these, it was trying to inverse tilemap.GetCellCenterWorld which would be able to return a tile's grid coordinates given world coordinates but thats a bit wierder on an isometric grid and for some reason, only unity will display code errors for me, not virtual studio. so i've given up on that functionality for the basic grid for now. I'm sure there are ways around it such as simply never storing grid coordinates.