unity project ready for AI
Download this assets folder, load it into the unity version, load samplescene from scenes, and you should be set up. This push is to begin development of the AI for each piece. Fundamentally, I don't care how you do it.
some stuff you need to know for testing stuff, on the tileholder object and on its script there is a list called "enable at start," in order to be placed on the grid at the moment a piece must be both set to inactive and on that list when you click play.
A bunch of values on the pieces, like starting position and time between turns, are set to a bunch of random values. We'll make decisions about these later, for now change them to whatever if you want to.
Remember to make any changes in the editor to the piece prefabs and not the specific objects in space.
Ideally, the only scripts that need to be edited to implement AI is the Turn() function in each piece's specific controller script.