Prerelease Mechanical Setup
A setup to allow the first zone to be a self-contained "game" with an end screen triggered by the trigger for the next zone, and changed the death setup to effectively restart the entire game added.  Includes 2 bugfixes involved with quitting to the main menu and loading back in, and one bugfix to do with the bishop AI. BEFORE WE SEND THIS OUT: someone needs to make sure UI is scaled correctly, and possibly rewrite the win text cuz what I have there works but could be better. Most likely, the only things that needs to be changed is the wintext, which is attached to the fadesystem -> fademenu canvas, needs to be rescaled to take up the top half of the screen, and the cleartext, thats on the display canvas, which can do whatever it wants but if you replace the object remember to reassign it as the clear text on the summer zonescript